refactor : NotificationDto.java
